NavMesh + Shader
Decided to mess more with NavMesh and the systems. I couldnt figure out why my baking wasnt detecting certain objects even with collider components in them. So I messed around with the NavMesh Volume Modifier and set the object to not walkable and rebaked.

For shader, I made it where the mushrooms have a glow effect to forage once you get one if you are in a close vicinity to them. I used a script machine to change the material.